+++ phpunit.xml
@@ -1,18 +1,31 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <phpunit xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:noNamespaceSchemaLocation="https://schema.phpunit.de/10.0/phpunit.xsd" backupGlobals="false" colors="true"
-         processIsolation="false" stopOnFailure="false" stopOnError="false" cacheDirectory=".phpunit.cache"
+         xsi:noNamespaceSchemaLocation="vendor/phpunit/phpunit/phpunit.xsd"
+         bootstrap="vendor/autoload.php"
+         backupGlobals="false"
+         colors="true"
+         processIsolation="false"
+         failOnRisky="false"
+         stopOnFailure="false"
+         stopOnError="false"
+         stderr="true"
+         cacheDirectory="runtime/.phpunit.cache"
          backupStaticProperties="false">
+    <coverage>
+        <report>
+            <html outputDirectory="build/coverage"/>
+            <text outputFile="build/coverage.txt"/>
+            <clover outputFile="build/logs/clover.xml"/>
+        </report>
+    </coverage>
     <testsuites>
-        <testsuite name="Package Tests">
-            <directory>tests/src</directory>
+        <testsuite name="Functional">
+            <directory>tests/Functional/src</directory>
+        </testsuite>
+        <testsuite name="Unit">
+            <directory>tests/Unit</directory>
         </testsuite>
     </testsuites>
-    <coverage>
-        <include>
-            <directory>src</directory>
-        </include>
-    </coverage>
     <php>
         <env name="MESSENGER_DEFAULT_PIPELINE" value="default"/>
         <env name="RR_MODE" value="jobs"/>
