+++ src/Context.php
@@ -7,13 +7,12 @@ namespace Spiral\Messenger;
 use Spiral\RoadRunner\Jobs\Task\ReceivedTaskInterface;
 use Symfony\Component\Messenger\Envelope;
 
-final class Context implements ContextInterface
+final readonly class Context implements ContextInterface
 {
     public function __construct(
-        private readonly Envelope $envelope,
-        private readonly ReceivedTaskInterface $task,
-    ) {
-    }
+        private Envelope $envelope,
+        private ReceivedTaskInterface $task,
+    ) {}
 
     public function getEnvelope(): Envelope
     {
